cleaned up soundplayer, unused variables, and non-gender-swap variables
this commit marks the beginning of the cleaning up of gender swap in a
different branch. The goal is to have it contained in a separate class
so that both setups can started from code in the same branch. Same goes
for sound player, it should be contained in its own class and given an
interface to load new sounds as wanted.